Store API: add shared cart-aware CheckoutRequirements capability
Introduce StoreApi\Utilities\CheckoutRequirements as the shared home for
'does this cart need field X' decisions that checkout consumers (POS, agentic,
payment gateways acting for a shopper) all need when they relax the default
field guards.

First rule: requires_email() is true when the cart has a downloadable line
item (it must be delivered somewhere), exposed through the
woocommerce_store_api_checkout_requires_email filter so product-type extensions
(gift cards, subscriptions, …) declare their own need once, shared across every
consumer, instead of each reimplementing the line-item walk.
